requirements.txt
pandas
polars
numpy
duckdb
statsmodels
scikit-learn
numba
joblib
plotly
streamlit
dash
great_expectations
pydantic
pydantic-settings
pyyaml
openpyxl
